+++ qty_calc/src/test/java/QtyCalcTests.java
@@ -73,159 +73,6 @@ public class QtyCalcTests {
         System.setIn(stdIn);
         System.setOut(stdOut);
     }    
-    ////// ITEM TESTS
-
-    @Test
-    public void itemAsJsonTest() {
-        Item i1 = new Item("thing1","9891",1000.0,"fl.oz");
-
-        JsonObject jo = i1.itemAsJson();
-
-        String name = (String) jo.get("name");
-        String dbid = (String) jo.get("ID");
-        String qty = (String) jo.get("quantity").toString();
-        String dbu = (String) jo.get("unit");
-
-        assertEquals("thing1",name);
-        assertEquals("9891",dbid);
-        assertEquals("1000.0",qty);
-        assertEquals("fl.oz",dbu);
-    }
-
-    @Test
-    public void Item_toString() {
-        Item i1 = new Item("thing1","9891",1000.0,"fl.oz");
-        Item i2 = new Item("thing1","9891",1000.0,"fl.oz");
-        Item i3 = new Item("thing3","1234",100.0,"liters");
-        assertTrue(i1.isEqual(i2));
-        assertFalse(i1.isEqual(i3));
-    }
-
-    @Test
-    public void Item_getters() {
-        Item i1 = new Item("thing1","9891",1000.0,"fl.oz");
-        assertEquals("thing1",i1.getItemName());
-        assertEquals("9891",i1.getDbID());
-        assertEquals("1000.0",i1.getQty().toString());
-        assertEquals("fl.oz",i1.getDbUnit());
-    }
-
-    /////// RECIPE TESTS
-
-    @Test
-    public void Recipe_isEqual_toString_tests() {
-        Recipe r1 = util.constructRecipe1();
-        Recipe r2 = util.constructRecipe1();
-        Recipe r3 = util.constructRecipe2();
-
-        assertEquals(r1.toString(),r2.toString());
-        assertNotEquals(r1.toString(),r3.toString());
-
-        assertTrue(r1.isEqual(r2));
-        assertFalse(r1.isEqual(r3));
-    }
-
-    @Test
-    public void addToRecipeTest_duplicate(){
-        ArrayList<Ingredient> thing1 = util.generateList1();
-        Recipe r1 = new Recipe("horrific mush", "0001", thing1);
-        Ingredient duplicate = new Ingredient("cheese", "5534", 8, "lbs");
-        boolean badAdd = r1.addIngredient(duplicate);
-        assertFalse(badAdd);
-    }
-
-    @Test
-    public void addToRecipeTest_badID(){
-        ArrayList<Ingredient> thing1 = util.generateList1();
-        Recipe r1 = new Recipe("horrific mush", "0001", thing1);
-        Ingredient badID = new Ingredient("dogfood", "5534", 8, "lbs");
-        boolean badAdd = r1.addIngredient(badID);
-        assertFalse(badAdd);
-    }
-
-    @Test
-    public void addToRecipeTest_good(){
-        ArrayList<Ingredient> thing1 = util.generateList1();
-        Recipe r1 = new Recipe("horrific mush", "0001", thing1);
-
-        String old = r1.toString();
-
-        Ingredient nuIng = new Ingredient("dogfood", "6789", 8, "lbs");
-        boolean goodAdd = r1.addIngredient(nuIng);
-        assertTrue(goodAdd);
-
-        String nu = r1.toString();
-        assertNotEquals(old, nu);
-    }
-
-    @Test
-    public void Recipe_remove_IngredientTests() {
-        Recipe r1 = util.constructRecipe1();
-
-        String old = r1.toString();
-
-        Ingredient nuIng = new Ingredient("dogfood", "6789", 8, "lbs");
-        assertFalse(r1.removeIngredient(nuIng));
-
-        Ingredient cheese = new Ingredient("cheese", "5534", 8, "lbs");
-        assertTrue(r1.removeIngredient(cheese));
-
-        String nu = r1.toString();
-        assertNotEquals(old, nu);
-    }
-
-    // tests if the constructed .json matches the originally created recipe
-    @Test
-    public void saveRecipeTest() throws Exception {
-
-        ArrayList<Ingredient> thing1 = util.generateList1();
-        Recipe r1 = new Recipe("test recipe", "1234", thing1);
-
-        if(BuildFile.RecipeExists("1234")){
-            setupIO();
-            String input = "y";
-            ByteArrayInputStream passIn = new ByteArrayInputStream(input.getBytes());
-            System.setIn(passIn);
-            r1.saveRecipeAsJson("recipe_folder");
-            restoreIO();
-        }
-        else {
-            r1.saveRecipeAsJson("recipe_folder");
-            assertTrue(BuildFile.RecipeExists("1234"));
-        }
-
-        String nuPath = "recipe_folder" + File.separator + "recipe_1234.json";
-        Recipe rjson = new Recipe(nuPath);
-        assertTrue(r1.isEqual(rjson));
-    }
-
-    @Test
-    public void testSaveJsonObject() throws IOException, JsonException {
-
-        ArrayList<Ingredient> thing1 = util.generateList1();
-        Recipe r1 = new Recipe("savejsontest", "0567", thing1);
-
-        JsonObject rjo = r1.getRecipeJson();
-
-        if(BuildFile.RecipeExists("0567")){
-            setupIO();
-            String input = "y";
-            ByteArrayInputStream passIn = new ByteArrayInputStream(input.getBytes());
-            System.setIn(passIn);
-            BuildFile.SaveRecipeJson(rjo, "recipe_folder");
-            restoreIO();
-        }
-        else {
-            BuildFile.SaveRecipeJson(rjo, "recipe_folder");
-        }
-
-        Recipe recipe_from_jo = new Recipe(rjo);
-        String rjost = r1.toString();
-        String jostr = recipe_from_jo.toString();
-
-        assertTrue(BuildFile.RecipeExists("0567"));
-        assertEquals(rjost,jostr);
-    }
 
 
     ///// ORDER TESTS
